requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pydicom>=2.4.0
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentencepiece>=0.1.99
tokenizers>=0.15.0
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1
matplotlib>=3.8.0
seaborn>=0.13.0
modal>=1.0.0
